union mask update
only uses regular method to look for first char in subquery, then shifts text, creates mask of remaining subquery characters and XNORs that with the text. Compares the result to some number of full set bytes to determine if the full subquery matches. should be somewhere in the range of O(m/sw + n/sw).